split file, add favicon, add unique id to each response
