+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 dist/
 .vscode/
 .parcel-cache/
+.env # This contains secrets, don't store in source control
