+++ package.json
@@ -22,13 +22,20 @@
   },
   "homepage": "https://github.com/romizka/numerical-deduction-app#readme",
   "devDependencies": {
+    "buffer": "^5.7.1",
+    "crypto-browserify": "^3.12.0",
+    "events": "^3.3.0",
+    "os-browserify": "^0.3.0",
     "parcel": "^2.8.3",
+    "path-browserify": "^1.0.1",
     "posthtml-doctype": "^1.1.1",
     "process": "^0.11.10",
+    "stream-browserify": "^3.0.0",
     "tailwindcss": "^3.3.1"
   },
   "dependencies": {
     "@splidejs/splide": "^4.1.4",
+    "dotenv": "^16.1.4",
     "firebase": "^9.22.1",
     "flowbite": "^1.6.4",
     "gh-pages": "^5.0.0",
