+++ public/index.html
@@ -9,7 +9,7 @@
     <link href="../src/assets/styles/index.css" rel="stylesheet" />
     <link href="https://cdnjs.cloudflare.com/ajax/libs/flowbite/1.6.4/flowbite.min.css" rel="stylesheet" />
     <link href="https://cdn.jsdelivr.net/npm/@splidejs/splide@4.1.4/dist/css/splide.min.css" rel="stylesheet">
-
+    <link href="/src/favicon.ico" rel="icon" type="image/x-icon" />
     <style>
