+++ src/.env
@@ -1,0 +1,1 @@
+API_KEY=AIzaSyAx8GpwgEZfIMkTItumcYm-M_XVeVWGxQI
